SPRINT VERSION DATE=August 3, 2010
1. Collected my working version routines that were not under subversion control
   into the RadAnl_Tools/sprint/trunk structure originally created by Dennis
   Flanigan and slightly modified by Chris Burghart.
2. Moved all original routines in the Sprint-src that did not agree with my
   working copies into Sprint-src/Original.  Many modifications were of a minor
   nature.
3. Moved all extra routines that were never being compiled in Sprint-src/Extra.
4. Unfortunately I had not used the svn move move command to complete #2 and #3.
   I set all moved routines in Sprint-src for deletion with "svn delete."
5. Only significant fix was a "SAVE IBEGRG" in routine NEXSWP.f to save the value
   of this variable from one CALL NEXSWP to the next.


git-svn-id: https://proxy.subversion.ucar.edu/RadAnl_Tools@17 569d80e6-d933-c0f4-fcc6-90685d5d544e

+++ sprint/trunk/Sprint-src/SUNFOR.c
@@ -1,5 +1,6 @@
 #include <stdio.h>
 #include "cedric.h"
+#include "op_sys.h"
 
 #define BYTSWRD 4
 
@@ -25,37 +26,44 @@ extern FILE *fp;  /* file pointer */
 
   /* read in first four bytes */
 
+  /* printf("+++ rdsunrec, rp=%d, ilen = %d, swapping= %d\n", *rp,*ilen,*swapping); */
   i = fread(&length, BYTSWRD, 1, fp);
   if( i <= 0 ) {
     *ilen = 0;
     return;
   }
+  /* printf("+++ rdsunrec: after first fread, i=%d, length=%d\n", i,length); */
 
   length = length >> (WORD_SIZE -32);
-  if(*swapping == 1) length = cswap32((long)length);
+  if(*swapping == 1) length = cswap32(length);
   if (length <= 0) {
     *ilen = 0;
     return;
   }
+  /* printf("+++ rdsunrec: after first fread, length=%d\n", length); */
   
   /* read in record */
 
   i = fread(rp, length, 1, fp);
+  /* printf("+++ rdsunrec: after second fread, i=%d\n", i); */
   if( i <= 0 ) {
     *ilen = 0;
     return;
   }
+  /* printf("+++ rdsunrec: after second fread, i=%d\n", i); */
 
   /* read tail of record */
 
   i = fread(reclength, BYTSWRD, 1, fp);
+  /* printf("+++ rdsunrec: after third fread, i=%d\n", i); */
   if( i <= 0 ) {
     *ilen = 0;
     return;
   }
-
+  /* printf("+++ rdsunrec: after third fread, i=%d\n", i); */
 
   *ilen = length;
+  /* printf("+++ rdsunrec: after third fread, ilen=%d\n", *ilen); */
 
   return;
